Added JS Docs for helper functions and seperated function out to another file
