+++ src/index.js
@@ -1,7 +1,9 @@
 import * as referenceHelpers from './helpers/referenceHelpers';
+import { doesReferenceContain } from './helpers/doesReferenceContain';
 export { referenceHelpers };
 export {
   cleanupReference,
   getVerses,
   parseReferenceToList,
- } from './helpers/referenceHelpers'
+ } from './helpers/referenceHelpers';
+export { doesReferenceContain };
